+++ doc/changelog.rst
@@ -10,6 +10,13 @@ New
 - Support LLVM 15
   (`#274 <https://github.com/bluescarni/heyoka/pull/274>`__).
 
+Fix
+~~~
+
+- Prevent callbacks from changing the time coordinate of the integrator.
+  This was never supported and could lead to crashes and/or hangs
+  in the ``propagate_*()`` functions
+  (`#278 <https://github.com/bluescarni/heyoka/pull/278>`__).
 
 0.19.0 (2022-09-18)
 -------------------